Refactors Telegram bot service and adds commands
This commit refactors the Telegram bot service by introducing command handlers for draw, reset, and mention functionalities.

This change enhances the modularity and maintainability of the bot's command processing logic.
